Move transportSettings() to a separate package
